requirements.txt
asgiref==3.5.0
Django==3.2.12
django-recaptcha==2.0.6
djangorestframework==3.13.1
mysqlclient==2.1.0
Pillow==9.0.0
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
django-widget-tweaks==1.4.12